11.5 A subquery can return only 1 column but we are returning all the columns from our union query. Let's just update that. Now let's run this query. We are all the way back to 450 ms.
What's happen. Let's run an explain to see if we can figure out. First we can see our query no longer use our last_name, first_name indexes. If we look at the select_type column we det a hint of what's wrong here. See how we have a dependence of query in two depending unions? That's not good. That means our new union query is not running in isolation, rather is being considered a dependency of the current query. And that's throwing everything off